refactor(source): consolidate bucket subpackage files and fetcher boilerplate
Three small cleanups that the deep-review audits called out:

- Fold pkg/source/bucket/endpoint.go and bucket/transport.go into
  bucket.go. Each was 39 LOC with one logical concern (Bucket
  endpoint normalization, TLS+proxy resolution) called only from
  bucket.go's Fetch. The split added three jumps to read one fetch
  path with no payoff. Test files stay independent.

- Drop ProxyConfig.HTTPProxyFunc — a single-call wrapper around URL()
  + http.ProxyURL. Inline the two lines into NewHTTPTransport's only
  caller site. URL() stays (covered by tests, useful as the typed
  proxy-resolution surface).

- Add source.ErrUnsupportedProvider. Each of git / oci / bucket
  carried a near-identical fmt.Errorf for the "provider %q is not
  implemented" gate; centralizing the shape keeps wording aligned as
  new providers land and gives operators one parseable error format
  across all source kinds.
